+++ tests/testthat/test_exceptions.R
@@ -149,7 +149,7 @@ test_that("pt2mod has no more than 2 elements", {
 
 test_that("Cannot add more than 100 patterns to a module", {
   expect_error({
-    mod$patterns[[101]] <- pt2_new_pattern()
+    mod$patterns[[101]] <- pt2_pattern()
   })
 })